Modularization components in nextjs-quill example
