+++ examples/nextjs-quill/types/index.ts
@@ -1,0 +1,10 @@
+import { Indexable, Text as YorkieText } from '@yorkie-js/sdk';
+
+export type YorkieDoc = {
+  content: YorkieText;
+};
+
+export type TextValueType = {
+  attributes?: Indexable;
+  content?: string;
+};
